userlogindone

+++ BackEnd/MaxFitGym/MaxFitGym/Models/RequestModel/MemberRegisterRequestDTO.cs
@@ -17,6 +17,6 @@
         public DateTime CreationDate { get; set; }
         public bool is_initalfeePaid { get; set; }
         public string Membershiptype { get; set; }
-        public decimal fee { get; set; }
+        public int fee { get; set; }
     }
 }
