+++ BackEnd/MaxFitGym/MaxFitGym/Models/RequestModel/MemberUpdateRequestDTO.cs
@@ -12,6 +12,6 @@
         public int Height { get; set; }
         public int Weight { get; set; }
         public string Membershiptype { get; set; }
-        public decimal fee { get; set; }
+        public int fee { get; set; }
     }
 }
